Shift to using an array of error messages
