+++ signin/internal/web/index.go
@@ -1,0 +1,27 @@
+package web
+
+import (
+	"html/template"
+	"log"
+	"net/http"
+)
+
+type Student struct {
+	Name       string
+	College    string
+	RollNumber int
+}
+
+func Index(w http.ResponseWriter, r *http.Request) {
+	student := Student{
+		Name:       "GB",
+		College:    "GolangBlogs",
+		RollNumber: 1,
+	}
+	parsedTemplate, _ := template.ParseFiles("Template/index.html")
+	err := parsedTemplate.Execute(w, student)
+	if err != nil {
+		log.Println("Error executing template :", err)
+		return
+	}
+}
